# swagger.yaml
openapi: 3.0.0
info:
  title: Akıllı Kütüphane Yönetim Sistemi API
  description: Flask ve MSSQL kullanılarak geliştirilmiş kütüphane yönetim sistemi REST API dokümantasyonu
  version: 1.0.0
  contact:
    name: API Support
    email: support@kutuphane.com
servers:
  - url: http://localhost:5000/api
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Book:
      type: object
      properties:
        kitapID:
          type: integer
        baslik:
          type: string
        isbn:
          type: string
        yayinYili:
          type: integer
        mevcutKopyaSayisi:
          type: integer
        toplamKopyaSayisi:
          type: integer
        kategoriID:
          type: integer
        kategoriAdi:
          type: string
        yayinEvi:
          type: string
        dil:
          type: string
        aciklama:
          type: string
...
    Author:
      type: object
      properties:
        yazarID:
          type: integer
        ad:
          type: string
        soyad:
          type: string
        dogumTarihi:
          type: string
          format: date
        ulke:
          type: string
        biyografi:
          type: string
...
paths:
...
  /books:
    get:
      tags:
        - Books
      summary: Tüm kitapları listele veya arama yap
      parameters:
        - name: title
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: integer
        - name: available_only
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Kitap listesi
    post:
      tags:
        - Books
      summary: Yeni kitap ekle (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Kitap eklendi
        '403':
          description: Yetkisiz erişim
  /books/{book_id}:
    get:
      tags:
        - Books
      summary: Kitap detayını getir
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Kitap detayı
    put:
      tags:
        - Books
      summary: Kitap güncelle (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '200':
          description: Kitap güncellendi
    delete:
      tags:
        - Books
      summary: Kitap sil (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Kitap silindi
...
  /authors:
    get:
      tags:
        - Authors
      summary: Tüm yazarları listele
      responses:
        '200':
          description: Yazar listesi
    post:
      tags:
        - Authors
      summary: Yeni yazar ekle (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '201':
          description: Yazar eklendi
  /authors/{author_id}:
    get:
      tags:
        - Authors
      summary: Yazar detayını getir
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Yazar detayı
    put:
      tags:
        - Authors
      summary: Yazar güncelle (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Author'
      responses:
        '200':
          description: Yazar güncellendi
    delete:
      tags:
        - Authors
      summary: Yazar sil (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: author_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Yazar silindi
...